last commit feb 12th

+++ models/employee_model.js
@@ -85,7 +85,7 @@ module.exports = (sequelize, dataTypes) => {
         unique: true,
       },
       earnings: {
-        type: dataTypes.INTEGER,
+        type: dataTypes.DECIMAL(10, 2),
       },
       deductions: {
         type: dataTypes.INTEGER,
